third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.secure.tentative.sub.http.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify no Partitioned Popins can be opened from insecure page
  assert_equals: expected "" but got "Secure-Insecure-PathOnly-"
Harness: the test ran to completion.
